tests/wpt/web-platform-tests/layout-instability/child-shift-with-parent.html

<!DOCTYPE html>
<title>Layout Instability: parent/child moved together</title>
<link rel="help" href="https://wicg.github.io/layout-instability/" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/util.js"></script>
<div id="parent" style="position: relative; width: 100px; height: 100px; border: 100px solid blue">
  <div id="child" style="height: 300px"></div>
</div>
<script>

promise_test(async () => {
  const watcher = new ScoreWatcher;

  // Wait for the initial render to complete.
  await waitForAnimationFrames(2);

  // Modify the position of the div.
  const parent = document.querySelector("#parent");
  parent.style.top = '100px';

  // If the implementation reports child and parent separately
  // (overlapping are should be excluded):
  const expectedScoreMin = computeExpectedScore(300 * (300 + 100) + 100 * 100, 100);
  // If the implementation reports parent bounding box (including child):
  const expectedScoreMax = computeExpectedScore(300 * (400 + 100), 100);

  // Observer fires after the frame is painted.
  assert_equals(watcher.score, 0);
  await watcher.promise;
  assert_between_inclusive(watcher.score, expectedScoreMin, expectedScoreMax);
}, 'Parent/child movement.');

</script>
